Split evm
