+++ packages/contracts/source/contracts/crypto/commit-handler.ts
@@ -1,0 +1,5 @@
+import type { ProcessableUnit } from "../processor/processable-unit.js";
+
+export interface CommitHandler {
+	onCommit(unit: ProcessableUnit): Promise<void>;
+}
