+++ packages/contracts/source/contracts/crypto/index.ts
@@ -1,5 +1,6 @@
 export * from "./block.js";
 export * from "./commit.js";
+export * from "./commit-handler.js";
 export * from "./config.js";
 export * from "./crypto.js";
 export * from "./enums.js";
