+++ packages/contracts/source/contracts/evm/index.ts
@@ -1,4 +1,5 @@
 export * from "./contract-service.js";
 export * from "./evm.js";
+export * from "./instance.js";
 export * from "./state.js";
 export * from "./worker.js";
